requirements.txt
asgiref==3.3.1
awsebcli==3.19.2
botocore==1.19.33
cement==2.8.2
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.3
dj-database-url==0.5.0
Django==3.1.4
django-environ==0.4.5
django-heroku==0.3.1
djangorestframework==3.12.2
future==0.16.0
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
pathspec==0.5.9
Pillow==8.0.1
psycopg2==2.8.6
pypiwin32==223
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.4
pywin32==300
PyYAML==5.3.1
requests==2.24.0
semantic-version==2.5.0
six==1.15.0
sqlparse==0.4.1
termcolor==1.1.0
urllib3==1.25.11
wcwidth==0.1.9
whitenoise==5.2.0
